+++ src/validate.test.ts
@@ -17,23 +17,34 @@ const bashScript = (async () => {
 });
 
 test('valid checksums', async () => {
-  const valid = await validateUploader(await bashScript());
+  const valid = validateUploader(await bashScript());
   expect(valid).toBeTruthy();
 });
 
 test('invalid checksums', async () => {
   const script = await bashScript();
-  const valid = await validateUploader(script.substring(0, script.length - 1));
+  const valid = validateUploader(script.substring(0, script.length - 1));
   expect(valid).toBeFalsy();
 });
 
 test('invalid script version', async () => {
   const script = await bashScript();
-  const valid = await validateUploader(script.substring(0, 20));
+  const valid = validateUploader(script.substring(0, 20));
   expect(valid).toBeFalsy();
 });
 
-test('invalid public checksum file', async () => {
-  const checksum = await retrieveChecksum('foo', 'bar');
+test('invalid public checksum file', () => {
+  const checksum = retrieveChecksum('foo', 'bar');
   expect(checksum).toBeFalsy();
 });
+
+test('invalid public checksum file', () => {
+  const checksum = retrieveChecksum('foo', 'bar');
+  expect(checksum).toBeFalsy();
+});
+
+test('invalid encryption', () => {
+  const checksum = retrieveChecksum('1.0.1', 'foo');
+  expect(checksum).toBeFalsy();
+});
+