+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 1.4.1
 ## Fixes
-- #287 Update VERSION regex to restrict on digits and dot
+- #287 Update VERSION regex to restrict on digits and dot and move checksums into script
 
 ## 1.4.0
 ### Features